4.15 - 4.23 done

+++ part4/controllers/blogs.js
@@ -1,8 +1,12 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable no-undef */
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request,response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username:1, name:1 })
   response.json(blogs)
 })
 
@@ -13,13 +17,37 @@ blogsRouter.get('/:id', async (request,response) => {
 
 
 blogsRouter.post('/', async (request,response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
 
-  const newBlog= await blog.save()
-  response.status(201).json(newBlog)
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+
+  const savedBlog= await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request,response) => {
+
+  const blogDetails = await Blog.findById(request.params.id)
+  const decodedUser = request.user
+
+  if(blogDetails.user.toString() !== decodedUser._id.toString() ){
+    return response.status(401).json({
+      error : 'The user does not have authority to delete '
+    })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
@@ -33,4 +61,6 @@ blogsRouter.put('/:id', async (request,response) => {
   response.status(201).json(updatedBlog)
 
 })
+
+
 module.exports = blogsRouter
